rename: board -> community
